Break BIRReader into smaller files
